requirements.txt
discord.py==2.3.2
aiohttp==3.10.5
python-dotenv==1.0.1
requests==2.32.3
async-timeout==4.0.3
colorama==0.4.6
logging==0.4.9.6
